Refactor OrderService for readability and functionality
- Refactored `OrderService` class for better readability and functionality.
- Added `IStockItemRepository` dependency to the constructor.
- Simplified `GetOrders` method using LINQ.
- Updated `GetOrderById` to throw `EntityNotFoundException` if not found and return `OrderDto`.
- Refactored `AddOrder` to use new private methods `GetClient` and `GetOrderItems`.
- Modified `UpdateOrder` to accept `orderId` parameter and update existing order.
- Enhanced `DeleteOrder` to check for order existence before deletion.
- Added private methods `GetOrderItems` and `GetClient` separating reusable logic to its own methods.
- Updated `IOrderService` interface to reflect `UpdateOrder` method signature change.

+++ orders_project_app/Service/Implementation/OrderService.cs
@@ -6,15 +6,15 @@ using orders_project_app.Service.Interface;
 
 namespace orders_project_app.Service.Implementation
 {
-    public class OrderService(IOrderRepository orderRepository, IClientRepository clientRepository) : IOrderService
+    public class OrderService(
+        IOrderRepository orderRepository,
+        IClientRepository clientRepository,
+        IStockItemRepository stockItemRepository) : IOrderService
     {
         public async Task<List<OrderDto>> GetOrders()
         {
             List<Order> orders = await orderRepository.GetOrders();
-            List<OrderDto> orderDtos = [];
-            orders.ForEach(order =>
-            {
-                OrderDto orderDto = new OrderDto
+            return orders.Select(order => new OrderDto
                 {
                     Id = order.Id,
                     OrderDateTime = order.OrderDateTime,
@@ -23,16 +23,13 @@ namespace orders_project_app.Service.Implementation
                     ClientId = order.ClientId,
                     Status = order.Status,
                     Type = order.Type
-                };
-                orderDtos.Add(orderDto);
-            });
-            return orderDtos;
+                }).ToList();
         }
 
         public async Task<OrderDto> GetOrderById(int id)
         {
-            Order orderEntity = await orderRepository.GetOrderById(id);
-            OrderDto orderDto = new OrderDto
+            Order orderEntity = await orderRepository.GetOrderById(id) ?? throw new EntityNotFoundException($"Order with ID='{id}' not found");
+            return new OrderDto
             {
                 Id = orderEntity.Id,
                 OrderDateTime = orderEntity.OrderDateTime,
@@ -42,51 +39,95 @@ namespace orders_project_app.Service.Implementation
                 Status = orderEntity.Status,
                 Type = orderEntity.Type
             };
-            return orderDto;
         }
 
         public async Task AddOrder(OrderCreateDto order)
         {
-            Client orderClient = await clientRepository.GetClientById(order.ClientId);
-            if (orderClient == null)
-            {
-                throw new EntityNotFoundException($"{typeof(Client)} with ID='{order.ClientId}' not found");
-            }
+            Client orderClient = await GetClient(order.ClientId);
+            List<OrderItem> orderItems = await GetOrderItems(order.OrderItems);
+            decimal totalPrice = orderItems.Sum(orderItem => orderItem.TotalPrice);
             Order orderEntity = new Order
             {
                 OrderDateTime = DateTime.Now,
-                OrderItems = order.OrderItems.Select(orderItemId => new OrderItem { Id = orderItemId }).ToList(), // This line is incorrect
-                TotalPrice = order.TotalPrice,
+                OrderItems = orderItems,
+                TotalPrice = totalPrice,
                 ClientId = order.ClientId,
                 Client = orderClient,
                 Status = order.Status,
                 Type = order.Type
             };
+
+            // Set order reference to orderItems
+            foreach (OrderItem orderItem in orderItems)
+            {
+                orderItem.Order = orderEntity;
+            }
+
+            //Persist order
             await orderRepository.AddOrder(orderEntity);
         }
 
-        public async Task UpdateOrder(OrderCreateDto order)
+        public async Task UpdateOrder(int orderId, OrderCreateDto order)
         {
-            Client orderClient = await clientRepository.GetClientById(order.ClientId);
-            if (orderClient == null)
+            var existingOrder = await orderRepository.GetOrderById(orderId)
+                                ?? throw new EntityNotFoundException($"Order with ID='{orderId}' not found");
+
+            existingOrder.Status = order.Status;
+            existingOrder.Type = order.Type;
+            var orderItems = await GetOrderItems(order.OrderItems);
+            existingOrder.TotalPrice = orderItems.Sum(o => o.TotalPrice);
+            existingOrder.OrderItems = orderItems;
+
+            foreach (var orderItem in orderItems)
             {
-                throw new EntityNotFoundException($"{typeof(Client)} with ID='{order.ClientId}' not found");
+                orderItem.Order = existingOrder;
             }
-            Order orderEntity = new Order
-            {
-                OrderItems = order.OrderItems.Select(orderItemId => new OrderItem { Id = orderItemId }).ToList(),
-                TotalPrice = order.TotalPrice,
-                ClientId = order.ClientId,
-                Client = orderClient,
-                Status = order.Status,
-                Type = order.Type
-            };
-            await orderRepository.UpdateOrder(orderEntity);
+
+            await orderRepository.UpdateOrder(existingOrder);
         }
 
         public async Task DeleteOrder(int id)
         {
+            var existingOrder = await orderRepository.GetOrderById(id);
+            if (existingOrder == null)
+            {
+                throw new EntityNotFoundException($"Order with ID='{id}' not found");
+            }
             await orderRepository.DeleteOrder(id);
+        }
+
+        private async Task<List<OrderItem>> GetOrderItems(List<OrderItemDto> orderItemDtos)
+        {
+            List<OrderItem> orderItems = [];
+
+            foreach (OrderItemDto orderItemDto in orderItemDtos)
+            {
+                StockItem stockItem = await stockItemRepository.GetStockItemById(orderItemDto.ItemId);
+                if (stockItem == null)
+                {
+                    throw new EntityNotFoundException($"{typeof(StockItem)} with ID='{orderItemDto.ItemId}' not found");
+                }
+                OrderItem orderItem = new OrderItem
+                {
+                    Id = stockItem.Id,
+                    Name = stockItem.Name,
+                    TotalPrice = stockItem.UnitPrice * orderItemDto.Quantity,
+                    Quantity = orderItemDto.Quantity
+                };
+                stockItem.Quantity -= orderItemDto.Quantity;
+                orderItems.Add(orderItem);
+            }
+            return orderItems;
+        }
+
+        private async Task<Client> GetClient(int clientId)
+        {
+            Client result = await clientRepository.GetClientById(clientId);
+            if (result == null)
+            {
+                throw new EntityNotFoundException($"{typeof(Client)} with ID='{clientId}' not found");
+            }
+            return result;
         }
     }
 }
